// apps/mms/src/components/pages/editor/containers/constraints/types.ts
export enum EventItem {
  ALL_LOCATIONS = 'ALL_LOCATIONS',
  ALL_USERS = 'ALL_USERS',
  ALWAYS = 'ALWAYS',
  SPECIFIC_DATE_TIME = 'SPECIFIC_DATE_TIME',
  SPECIFIC_EVENT_STATES = 'SPECIFIC_EVENT_STATES',
  SPECIFIC_LOCATIONS = 'SPECIFIC_LOCATIONS',
  SPECIFIC_STATES = 'SPECIFIC_STATES',
  SPECIFIC_USERS = 'SPECIFIC_USERS',
}